% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscNoise.R
\name{build_model}
\alias{build_model}
\title{This is an internal function that builds the classifier model}
\usage{
build_model(classifier, train, ...)
}
\arguments{
\item{classifier}{a string, takes the name of the classifier.Currently supported classifiers are 
'rf' - Random forest
'lrm' - Logistic regression
'CART' - Classification tree
'knn' - K-Nearest Neighbors}

\item{train}{a dataframe. Contains the train dataset}

\item{...}{optional additional arguements for the classifer}
}
\description{
This is an internal function that builds the classifier model
}
