% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscNoise.R
\name{create_oosb_sample}
\alias{create_oosb_sample}
\title{The function generates the out-of-sample bootstrap train and test data
A function to compute Area Under the ROC-CURVE}
\usage{
create_oosb_sample(data, boot_size, seed = TRUE)
}
\arguments{
\item{data}{must be a object of type data.frame, with the continuous dependent variable}

\item{boot_size}{a numeric value. It specifies the number of bootstrap iterations to be used in the framework.}

\item{seed}{a logical value indicating if the experiment needs to repetable or random. Defaults to TRUE}
}
\value{
a list containing train and test data for all specified boot_Size iterations
}
\description{
The function generates the out-of-sample bootstrap train and test data
A function to compute Area Under the ROC-CURVE
}
