% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscNoise.R
\name{get_auc}
\alias{get_auc}
\title{Extract AUC Function
A function to compute Area Under the ROC-CURVE}
\usage{
get_auc(actuals, predicted)
}
\arguments{
\item{actuals}{A vector containing the true class of the test data points}

\item{predicted}{A vector of predicted values}
}
\value{
a numeric copmuted MCC value
}
\description{
Extract AUC Function
A function to compute Area Under the ROC-CURVE
}
