% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscNoise.R
\name{get_mcc}
\alias{get_mcc}
\title{A function to compute Mathew's Correlation Coefficient}
\usage{
get_mcc(act, pred, classes)
}
\arguments{
\item{act}{A vector containing the true class of the test data points}

\item{pred}{A vector of predicted values}

\item{classes}{The classes that are existent in the dependent variable}
}
\value{
a numeric copmuted MCC value
}
\description{
A function to compute Mathew's Correlation Coefficient
}
