% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscNoise.R
\name{get_performance_metrics}
\alias{get_performance_metrics}
\title{The function that is responsible for computing the performance evaluation metrics}
\usage{
get_performance_metrics(actuals, test, model)
}
\arguments{
\item{actuals}{A vector containing the true class of the test data points}

\item{test}{a dataframe. Test dataset}

\item{model}{a r model. A trained classifier model whose performance is to be evaluated}
}
\value{
a vector with al the performance measures computed
}
\description{
The function that is responsible for computing the performance evaluation metrics
}
