% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscNoise.R
\name{predict_generic}
\alias{predict_generic}
\title{Abstracting the predict function}
\usage{
predict_generic(model, test, type)
}
\arguments{
\item{model}{r model object, passed to generate actual responses or probability scores}

\item{test}{the test dataset for which prob score or responses are to be genereated}

\item{type}{a string, that takes either c('class','prob','response') for apporiate response}
}
\value{
probability score for the outcome classes or actual class that the supplied model predicts
}
\description{
Abstracting the predict function
}
