% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscNoise.R
\name{RWKH_framework}
\alias{RWKH_framework}
\title{RWKH framework
This framework does the heavy lifting of computing the performance and featue importance using out-o-sample boostrap validation for a given data configuration}
\usage{
RWKH_framework(classifier, data, parallel, n_cores, boot_size, dep_var,
  cutpoint, target, ...)
}
\arguments{
\item{classifier}{a string, takes the name of the classifier.Currently supported classifiers are 
'rf' - Random forest
'lrm' - Logistic regression
'CART' - Classification tree
'knn' - K-Nearest Neighbors}

\item{data}{must be a object of type data.frame, with the continuous dependent variable}

\item{parallel}{a logical value indicating if the function must be executed in parallel --Recommended.}

\item{n_cores}{a numeric value specifying the number of cores to be used for parallel execution. Defaults to 1.}

\item{boot_size}{a numeric value. It specifies the number of bootstrap iterations to be used in the framework. Defaults to 100}

\item{dep_var}{a string giving the column name of continuous dependent variable supplied in the data parameter. This is the variable which creates the discretization noise.}

\item{target}{a numeric value indicating the amount of discretization noise is to be included relative to cutpoint'}
}
\value{
a list, that contains two lists containing performance impacts and importance values for each bootstrap iteration for the classifier
}
\description{
RWKH framework
This framework does the heavy lifting of computing the performance and featue importance using out-o-sample boostrap validation for a given data configuration
}
